fix(native-ros2): resolve transport bugs across all sensor modules
- DepthCamera: invoke OnCameraDataGenerated/OnCameraInfoGenerated events
  in ImageProcessing() override (was silently dropping native ROS2 data)
- SegmentationCamera: invoke parent Camera events alongside
  OnSegmentationDataGenerated for native image publishing
- PointCloud2Struct: add missing point_step field to match C++ layout
  (fixes 3D LiDAR data corruption from struct misalignment)
- MicomPlugin: fix odometry quaternion from yaw via sin/cos(yaw/2)
  instead of incorrectly using angular velocity Z as quaternion Z
- ContactPlugins: update for ROS2 Jazzy API (bindings→contacts,
  collision fields use Entity type with .name accessor)
- cloisim_ros2.cpp: add JointState C++ publisher + fix Contacts API

Validated: IMU (93 Hz), Odometry (47 Hz) confirmed working via benchmark.
Camera/LiDAR 0 Hz is a preexisting Unity AsyncGPUReadback issue
affecting both native and ZMQ modes equally.

+++ Assets/Scripts/CLOiSimPlugins/SegmentationCameraPlugin.cs
@@ -75,7 +75,7 @@ public class SegmentationCameraPlugin : CameraPlugin
 		
 		for (int i = 0; i < numLabels; i++)
 		{
-			classIds[i] = msg.ClassMaps[i].ClassId;
+			classIds[i] = (int)msg.ClassMaps[i].ClassId;
 			classNames[i] = msg.ClassMaps[i].ClassName;
 		}
